# gen/http/openapi3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fire Tracker Service",
    "description": "Service for fire tracking",
    "termsOfService": "terms",
    "contact": {
      "name": "support",
      "url": "firetracker.com",
      "email": "support@firetracker.com"
    },
    "license": {
      "name": "GPT-3",
      "url": "licence url"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://test.tmf.com"
    },
    {
      "url": "https://trackmyfire.com"
    }
  ],
  "paths": {
...
    "/fire/{id}/logs": {
      "get": {
        "tags": [
          "fire"
        ],
        "summary": "getLogsForFire fire",
        "description": "Gets a list of logs for a fire",
        "operationId": "fire#getLogsForFire",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "fire id",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "fire id",
              "example": 1841298947812401949,
              "format": "int64"
            },
            "example": 204222991974978568
          }
        ],
        "responses": {
          "200": {
            "description": "OK response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogList"
                },
                "example": {
                  "logs": [
                    {
                      "addedAt": "2020-01-02T00:00:00Z",
                      "createdAt": "2020-01-02T00:00:00Z",
                      "fireID": 1,
                      "id": 5636220753615159119,
                      "name": "Big Log",
                      "size": "L",
                      "updatedAt": "2020-01-02T00:00:00Z",
                      "weather": {
                        "createdAt": "2020-01-02T00:00:00Z",
                        "dewPoint": {
                          "unit": "C",
                          "value": 20
                        },
                        "fireID": 1,
                        "humidity": 1500,
                        "id": 1,
                        "logID": 1,
                        "temperature": {
                          "unit": "C",
                          "value": 20
                        },
                        "weatherTime": "2020-01-02T00:00:00Z",
                        "weatherType": "Sunny",
                        "wind": {
                          "direction": "E",
                          "speed": 15,
                          "unit": "MPH"
                        }
                      }
                    },
                    {
                      "addedAt": "2020-01-02T00:00:00Z",
                      "createdAt": "2020-01-02T00:00:00Z",
                      "fireID": 1,
                      "id": 5636220753615159119,
                      "name": "Big Log",
                      "size": "L",
                      "updatedAt": "2020-01-02T00:00:00Z",
                      "weather": {
                        "createdAt": "2020-01-02T00:00:00Z",
                        "dewPoint": {
                          "unit": "C",
                          "value": 20
                        },
                        "fireID": 1,
                        "humidity": 1500,
                        "id": 1,
                        "logID": 1,
                        "temperature": {
                          "unit": "C",
                          "value": 20
                        },
                        "weatherTime": "2020-01-02T00:00:00Z",
                        "weatherType": "Sunny",
                        "wind": {
                          "direction": "E",
                          "speed": 15,
                          "unit": "MPH"
                        }
                      }
                    },
                    {
                      "addedAt": "2020-01-02T00:00:00Z",
                      "createdAt": "2020-01-02T00:00:00Z",
                      "fireID": 1,
                      "id": 5636220753615159119,
                      "name": "Big Log",
                      "size": "L",
                      "updatedAt": "2020-01-02T00:00:00Z",
                      "weather": {
                        "createdAt": "2020-01-02T00:00:00Z",
                        "dewPoint": {
                          "unit": "C",
                          "value": 20
                        },
                        "fireID": 1,
                        "humidity": 1500,
                        "id": 1,
                        "logID": 1,
                        "temperature": {
                          "unit": "C",
                          "value": 20
                        },
                        "weatherTime": "2020-01-02T00:00:00Z",
                        "weatherType": "Sunny",
                        "wind": {
                          "direction": "E",
                          "speed": 15,
                          "unit": "MPH"
                        }
                      }
                    },
                    {
                      "addedAt": "2020-01-02T00:00:00Z",
                      "createdAt": "2020-01-02T00:00:00Z",
                      "fireID": 1,
                      "id": 5636220753615159119,
                      "name": "Big Log",
                      "size": "L",
                      "updatedAt": "2020-01-02T00:00:00Z",
                      "weather": {
                        "createdAt": "2020-01-02T00:00:00Z",
                        "dewPoint": {
                          "unit": "C",
                          "value": 20
                        },
                        "fireID": 1,
                        "humidity": 1500,
                        "id": 1,
                        "logID": 1,
                        "temperature": {
                          "unit": "C",
                          "value": 20
                        },
                        "weatherTime": "2020-01-02T00:00:00Z",
                        "weatherType": "Sunny",
                        "wind": {
                          "direction": "E",
                          "speed": 15,
                          "unit": "MPH"
                        }
                      }
                    }
                  ],
                  "pagination": {
                    "limit": 10,
                    "page": 1,
                    "total": 15
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "id": "3F1FKVRR",
                  "message": "Value of ID must be an integer",
                  "name": "bad_request"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogList"
                },
                "example": {
                  "logs": [
                    {
                      "addedAt": "2020-01-02T00:00:00Z",
                      "createdAt": "2020-01-02T00:00:00Z",
                      "fireID": 1,
                      "id": 5636220753615159119,
                      "name": "Big Log",
                      "size": "L",
                      "updatedAt": "2020-01-02T00:00:00Z",
                      "weather": {
                        "createdAt": "2020-01-02T00:00:00Z",
                        "dewPoint": {
                          "unit": "C",
                          "value": 20
                        },
                        "fireID": 1,
                        "humidity": 1500,
                        "id": 1,
                        "logID": 1,
                        "temperature": {
                          "unit": "C",
                          "value": 20
                        },
                        "weatherTime": "2020-01-02T00:00:00Z",
                        "weatherType": "Sunny",
                        "wind": {
                          "direction": "E",
                          "speed": 15,
                          "unit": "MPH"
                        }
                      }
                    },
                    {
                      "addedAt": "2020-01-02T00:00:00Z",
                      "createdAt": "2020-01-02T00:00:00Z",
                      "fireID": 1,
                      "id": 5636220753615159119,
                      "name": "Big Log",
                      "size": "L",
                      "updatedAt": "2020-01-02T00:00:00Z",
                      "weather": {
                        "createdAt": "2020-01-02T00:00:00Z",
                        "dewPoint": {
                          "unit": "C",
                          "value": 20
                        },
                        "fireID": 1,
                        "humidity": 1500,
                        "id": 1,
                        "logID": 1,
                        "temperature": {
                          "unit": "C",
                          "value": 20
                        },
                        "weatherTime": "2020-01-02T00:00:00Z",
                        "weatherType": "Sunny",
                        "wind": {
                          "direction": "E",
                          "speed": 15,
                          "unit": "MPH"
                        }
                      }
                    },
                    {
                      "addedAt": "2020-01-02T00:00:00Z",
                      "createdAt": "2020-01-02T00:00:00Z",
                      "fireID": 1,
                      "id": 5636220753615159119,
                      "name": "Big Log",
                      "size": "L",
                      "updatedAt": "2020-01-02T00:00:00Z",
                      "weather": {
                        "createdAt": "2020-01-02T00:00:00Z",
                        "dewPoint": {
                          "unit": "C",
                          "value": 20
                        },
                        "fireID": 1,
                        "humidity": 1500,
                        "id": 1,
                        "logID": 1,
                        "temperature": {
                          "unit": "C",
                          "value": 20
                        },
                        "weatherTime": "2020-01-02T00:00:00Z",
                        "weatherType": "Sunny",
                        "wind": {
                          "direction": "E",
                          "speed": 15,
                          "unit": "MPH"
                        }
                      }
                    }
                  ],
                  "pagination": {
                    "limit": 10,
                    "page": 1,
                    "total": 15
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "fault": {
            "type": "boolean",
            "description": "Is the error a server-side fault?",
            "example": false
          },
          "id": {
            "type": "string",
            "description": "ID is a unique identifier for this particular occurrence of the problem.",
            "example": "123abc"
          },
          "message": {
            "type": "string",
            "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
            "example": "parameter 'p' must be an integer"
          },
          "name": {
            "type": "string",
            "description": "Name is the name of this class of errors.",
            "example": "bad_request"
          },
          "temporary": {
            "type": "boolean",
            "description": "Is the error temporary?",
            "example": true
          },
          "timeout": {
            "type": "boolean",
            "description": "Is the error a timeout?",
            "example": true
          }
        },
        "description": "Please ensure you are sending valid data.",
        "example": {
          "id": "3F1FKVRR",
          "message": "Value of ID must be an integer",
          "name": "bad_request"
        },
        "required": [
          "name",
          "id",
          "message",
          "temporary",
          "timeout",
          "fault"
        ]
      },
...
      "LogList": {
        "type": "object",
        "properties": {
          "logs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateRequestBody"
            },
            "description": "logs",
            "example": [
              {
                "addedAt": "2020-01-02T00:00:00Z",
                "createdAt": "2020-01-02T00:00:00Z",
                "fireID": 1,
                "id": 5636220753615159119,
                "name": "Big Log",
                "size": "L",
                "updatedAt": "2020-01-02T00:00:00Z",
                "weather": {
                  "createdAt": "2020-01-02T00:00:00Z",
                  "dewPoint": {
                    "unit": "C",
                    "value": 20
                  },
                  "fireID": 1,
                  "humidity": 1500,
                  "id": 1,
                  "logID": 1,
                  "temperature": {
                    "unit": "C",
                    "value": 20
                  },
                  "weatherTime": "2020-01-02T00:00:00Z",
                  "weatherType": "Sunny",
                  "wind": {
                    "direction": "E",
                    "speed": 15,
                    "unit": "MPH"
                  }
                }
              },
              {
                "addedAt": "2020-01-02T00:00:00Z",
                "createdAt": "2020-01-02T00:00:00Z",
                "fireID": 1,
                "id": 5636220753615159119,
                "name": "Big Log",
                "size": "L",
                "updatedAt": "2020-01-02T00:00:00Z",
                "weather": {
                  "createdAt": "2020-01-02T00:00:00Z",
                  "dewPoint": {
                    "unit": "C",
                    "value": 20
                  },
                  "fireID": 1,
                  "humidity": 1500,
                  "id": 1,
                  "logID": 1,
                  "temperature": {
                    "unit": "C",
                    "value": 20
                  },
                  "weatherTime": "2020-01-02T00:00:00Z",
                  "weatherType": "Sunny",
                  "wind": {
                    "direction": "E",
                    "speed": 15,
                    "unit": "MPH"
                  }
                }
              }
            ]
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "example": {
          "logs": [
            {
              "addedAt": "2020-01-02T00:00:00Z",
              "createdAt": "2020-01-02T00:00:00Z",
              "fireID": 1,
              "id": 5636220753615159119,
              "name": "Big Log",
              "size": "L",
              "updatedAt": "2020-01-02T00:00:00Z",
              "weather": {
                "createdAt": "2020-01-02T00:00:00Z",
                "dewPoint": {
                  "unit": "C",
                  "value": 20
                },
                "fireID": 1,
                "humidity": 1500,
                "id": 1,
                "logID": 1,
                "temperature": {
                  "unit": "C",
                  "value": 20
                },
                "weatherTime": "2020-01-02T00:00:00Z",
                "weatherType": "Sunny",
                "wind": {
                  "direction": "E",
                  "speed": 15,
                  "unit": "MPH"
                }
              }
            },
            {
              "addedAt": "2020-01-02T00:00:00Z",
              "createdAt": "2020-01-02T00:00:00Z",
              "fireID": 1,
              "id": 5636220753615159119,
              "name": "Big Log",
              "size": "L",
              "updatedAt": "2020-01-02T00:00:00Z",
              "weather": {
                "createdAt": "2020-01-02T00:00:00Z",
                "dewPoint": {
                  "unit": "C",
                  "value": 20
                },
                "fireID": 1,
                "humidity": 1500,
                "id": 1,
                "logID": 1,
                "temperature": {
                  "unit": "C",
                  "value": 20
                },
                "weatherTime": "2020-01-02T00:00:00Z",
                "weatherType": "Sunny",
                "wind": {
                  "direction": "E",
                  "speed": 15,
                  "unit": "MPH"
                }
              }
            }
          ],
          "pagination": {
            "limit": 10,
            "page": 1,
            "total": 15
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "integer",
            "description": "max number of things",
            "example": 10,
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "description": "page number",
            "example": 1,
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "description": "count of the things",
            "example": 15,
            "format": "int64"
          }
        },
        "example": {
          "limit": 10,
          "page": 1,
          "total": 15
        }
      },
...
    }
  }
}